+++ src/TestSuite/TestCase.php
@@ -18,8 +18,6 @@ namespace MeTools\TestSuite;
 use Cake\Core\Configure;
 use Cake\ORM\Table;
 use Cake\TestSuite\TestCase as CakeTestCase;
-use Throwable;
-use Tools\Exceptionist;
 use Tools\Filesystem;
 use Tools\TestSuite\TestTrait;
 
@@ -90,9 +88,7 @@ abstract class TestCase extends CakeTestCase
     {
         parent::tearDownAfterClass();
 
-        if (LOGS !== TMP) {
-            Filesystem::unlinkRecursive(LOGS, ['.gitkeep', 'empty'], true);
-        }
+        Filesystem::unlinkRecursive(LOGS, '.gitkeep', true);
     }
 
     /**
@@ -104,14 +100,8 @@ abstract class TestCase extends CakeTestCase
      */
     public function assertLogContains(string $expectedContent, string $filename, string $message = ''): void
     {
-        try {
-            $filename = $this->getLogFullPath($filename);
-            $content = file_get_contents(Exceptionist::isReadable($filename)) ?: '';
-        } catch (Throwable $e) {
-            $this->fail($e->getMessage());
-        }
-
-        $this->assertStringContainsString($expectedContent, $content, $message);
+        $this->assertFileIsReadable($filename);
+        $this->assertStringContainsString($expectedContent, file_get_contents($filename) ?: '', $message);
     }
 
     /**
@@ -121,10 +111,14 @@ abstract class TestCase extends CakeTestCase
      * @param string $message The failure message that will be appended to the generated message
      * @return void
      * @since 2.20.7
+     * @deprecated 2.24.1 Use instead `assertStringEndsNotWith()`
+     * @codeCoverageIgnore
      */
     protected function assertSqlEndsNotWith(string $suffix, string $sql, string $message = ''): void
     {
-        $this->assertStringEndsNotWith(str_replace('`', '', $suffix), str_replace('`', '', $sql), $message);
+        deprecationWarning('Deprecated. Use instead `assertStringEndsNotWith()`');
+
+        $this->assertStringEndsNotWith($suffix, $sql, $message);
     }
 
     /**
@@ -134,19 +128,27 @@ abstract class TestCase extends CakeTestCase
      * @param string $message The failure message that will be appended to the generated message
      * @return void
      * @since 2.20.7
+     * @deprecated 2.24.1 Use instead `assertStringEndsWith()`
+     * @codeCoverageIgnore
      */
     protected function assertSqlEndsWith(string $suffix, string $sql, string $message = ''): void
     {
-        $this->assertStringEndsWith(str_replace('`', '', $suffix), str_replace('`', '', $sql), $message);
+        deprecationWarning('Deprecated. Use instead `assertStringEndsWith()`');
+
+        $this->assertStringEndsWith($suffix, $sql, $message);
     }
 
     /**
      * Deletes a log file
      * @param string $filename Log filename
      * @return void
+     * @deprecated 2.24.1
+     * @codeCoverageIgnore
      */
     public function deleteLog(string $filename): void
     {
+        deprecationWarning('`TestCase::deleteLog()` is deprecated and will be removed in a later release');
+
         unlink($this->getLogFullPath($filename));
     }
 
@@ -155,12 +157,14 @@ abstract class TestCase extends CakeTestCase
      * @param string $filename Log filename
      * @return string
      * @since 2.16.10
+     * @deprecated 2.24.1
+     * @codeCoverageIgnore
      */
     protected function getLogFullPath(string $filename): string
     {
-        $filename .= Filesystem::getExtension($filename) ? '' : '.log';
+        deprecationWarning('`TestCase::getLogFullPath()` is deprecated and will be removed in a later release');
 
-        return Filesystem::makePathAbsolute($filename, LOGS);
+        return Filesystem::makePathAbsolute($filename . Filesystem::getExtension($filename) ? '' : '.log', LOGS);
     }
 
     /**